Mac Carbon 3.8.11b1 	Eradicate most usages of FSSpec, except for drop file logic (sigh) and one in InternetConfig Remove getFileNameByFSSpec logic Alter getFileNameByFSRef to drop items needed by getFileNameByFSSpec. Reworked much of drag and drop logic
git-svn-id: http://squeakvm.org/svn/squeak/trunk@1348 fa1542d4-bde8-0310-ad64-8ed1123d492a

+++ platforms/Mac OS/vm/sqMacMain.c
@@ -321,7 +321,7 @@ char * GetAttributeString(int id) {
 		static char data[32];
 		
 		Gestalt(gestaltSystemVersion, &myattr);
-		sprintf(data,"%X",myattr);
+		sprintf(data,"%X",(unsigned int) myattr);
 		return data;
 	}
 	if (id == 1003) {
